main-backend-service/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
collection==0.1.6
cryptography==46.0.3
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.11
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
joblib==1.5.2
multidict==6.7.0
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
postgrest==2.22.2
propcache==0.4.1
psycopg==3.2.12
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.22.2
regex==2025.10.23
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.22.2
StrEnum==0.4.15
supabase==2.22.2
supabase-auth==2.22.2
supabase-functions==2.22.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
websockets==15.0.1
yarl==1.22.0

requests
openai
